UI优化，部分代码细节改动

+++ app/src/main/java/com/example/filemanager/tools/FileDeleter.java
@@ -2,10 +2,16 @@ package com.example.filemanager.tools;
 
 import java.io.File;
 
+//FileDeleter类封装了与删除有关的静态方法，可通过类名直接调用
 public class FileDeleter {
+    /*
+    全部删除方法
+    如果file是文件，直接删除该文件
+    如果file是目录，判断目录是否为空
+    空目录直接删除，非空目录递归调用本方法
+     */
     public static boolean deleteAll(File file) {
         boolean isDeleted = false;
-
         if (file.isFile()) {
             file.delete();
             isDeleted = true;
@@ -25,16 +31,24 @@ public class FileDeleter {
         return isDeleted;
     }
 
+    /*
+    清除内部文件方法
+    dir是目录时，判断dir是否为空
+    dir内部为空时不用删除
+    dir内部非空时对内部每一个file调用deleteAll()方法
+     */
     public static boolean deleteInner(File dir) {
-        boolean isDeleted=false;
-        File[] files = dir.listFiles();
-        if (files == null) {
-            isDeleted= true;
-        } else {
-            for (File file : files) {
-                deleteAll(file);
+        boolean isDeleted = false;
+        if (dir.isDirectory()) {
+            File[] files = dir.listFiles();
+            if (files == null) {
+                isDeleted = true;
+            } else {
+                for (File file : files) {
+                    deleteAll(file);
+                }
+                isDeleted = true;
             }
-            isDeleted= true;
         }
         return isDeleted;
     }
